requirements.txt
streamlit==1.32.0
Pillow==10.2.0
numpy==1.26.4
opencv-python==4.9.0.80
rembg==2.0.50 